Separated accortion items into new files
